multitenant: move tenantcostmodel
Move `tenantcostmodel` from `pkg/util` to the new `pkg/multitenant`.

Release note: None
