+++ pkg/kv/kvserver/tenantrate/BUILD.bazel
@@ -14,13 +14,13 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/multitenant",
+        "//pkg/multitenant/tenantcostmodel",
         "//pkg/roachpb:with-mocks",
         "//pkg/settings",
         "//pkg/util/metric",
         "//pkg/util/metric/aggmetric",
         "//pkg/util/quotapool",
         "//pkg/util/syncutil",
-        "//pkg/util/tenantcostmodel",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
@@ -33,13 +33,13 @@ go_test(
     data = glob(["testdata/**"]),
     deps = [
         ":tenantrate",
+        "//pkg/multitenant/tenantcostmodel",
         "//pkg/roachpb:with-mocks",
         "//pkg/settings/cluster",
         "//pkg/testutils",
         "//pkg/testutils/metrictestutils",
         "//pkg/util/leaktest",
         "//pkg/util/metric",
-        "//pkg/util/tenantcostmodel",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_errors//:errors",
