+++ pkg/multitenant/tenantcostmodel/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "model.go",
         "settings.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/util/tenantcostmodel",
+    importpath = "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcostmodel",
     visibility = ["//visibility:public"],
     deps = ["//pkg/settings"],
 )
